+++ playground/stimulus/assets/controllers/counter_controller.scss
