+++ playground/stimulus/assets/controllers/welcome_controller.ts
@@ -1,6 +1,12 @@
 import { Controller } from "@hotwired/stimulus"
 
+import.meta.stimulusFetch = "eager";
+import.meta.stimulusControllerIdentifier = "welcome";
+
 export default class controller extends Controller {
+  declare titleTarget: HTMLDivElement;
+  declare nameValue: string;
+
   static targets = [
     "title",
   ]
