+++ playground/stimulus/assets/page/translator/index.ts
@@ -5,27 +5,27 @@ import {
   trans,
 } from './translator';
 
-document.getElementById('hello').innerText = trans(HELLO);
-document.getElementById('say_hello').innerText = trans(SAY_HELLO, {
+document.getElementById('hello')!.innerText = trans(HELLO);
+document.getElementById('say_hello')!.innerText = trans(SAY_HELLO, {
   name: 'Fabien'
 });
-document.getElementById('invitation_title').innerText = trans(INVITATION_TITLE, {
+document.getElementById('invitation_title')!.innerText = trans(INVITATION_TITLE, {
   organizer_gender: 'female',
   organizer_name: 'Clara',
 });
-document.getElementById('num_of_apples').innerText = trans(NUM_OF_APPLES, {
+document.getElementById('num_of_apples')!.innerText = trans(NUM_OF_APPLES, {
   apples: 2
 });
-document.getElementById('finish_place').innerText = trans(FINISH_PLACE, {
+document.getElementById('finish_place')!.innerText = trans(FINISH_PLACE, {
   place: 1
 });
-document.getElementById('published_at').innerText = trans(PUBLISHED_AT, {
+document.getElementById('published_at')!.innerText = trans(PUBLISHED_AT, {
   publication_date: new Date('2023-04-27T08:12')
 });
-document.getElementById('progress').innerText = trans(PROGRESS, {
+document.getElementById('progress')!.innerText = trans(PROGRESS, {
   progress: 0.58,
 });
-document.getElementById('value_of_object').innerText = trans(VALUE_OF_OBJECT, {
+document.getElementById('value_of_object')!.innerText = trans(VALUE_OF_OBJECT, {
   value: 30
 });
 // No parameters, uses the default domain ("messages") and the default locale
