+++ playground/stimulus/assets/react/controllers/App.tsx
