+++ src/vite-plugin-symfony/src/stimulus/helpers/index.ts
@@ -38,12 +38,20 @@ export function startStimulusApp() {
   app.debug = process.env.NODE_ENV === "development";
 
   for (const controllerInfos of thirdPartyControllers) {
-    console.log(`register Stimulus controller : ${controllerInfos.identifier} (${controllerInfos.fetch})`);
     if (controllerInfos.fetch === "lazy") {
       app.register(controllerInfos.identifier, createLazyController(controllerInfos.controller));
     } else {
       app.register(controllerInfos.identifier, controllerInfos.controller);
     }
+  }
+
+  if (app.debug) {
+    console.groupCollapsed("application register controllers from controllers.json");
+    console.log(
+      "controllers",
+      thirdPartyControllers.map((infos) => infos.identifier),
+    );
+    console.groupEnd();
   }
 
   return app;
@@ -83,6 +91,8 @@ function isStimulusControllerInfosImport(
 }
 
 export function registerControllers(app: Application, modules: Modules) {
+  const controllersAdded: string[] = [];
+
   Object.entries(modules).forEach(([filePath, unknownController]) => {
     const identifier = getStimulusControllerId(filePath);
     if (!identifier) {
@@ -90,19 +100,28 @@ export function registerControllers(app: Application, modules: Modules) {
     }
     if (isLazyLoadedControllerModule(unknownController)) {
       app.register(identifier, createLazyController(unknownController));
+      controllersAdded.push(identifier);
     } else if (isStimulusControllerConstructor(unknownController)) {
       app.register(identifier, unknownController);
+      controllersAdded.push(identifier);
     } else if (isStimulusControllerInfosImport(unknownController)) {
-      registerController(app, unknownController.default);
+      registerController(app, unknownController.default, false);
+      controllersAdded.push(unknownController.default.identifier);
     } else {
       throw new Error(
         `unknown Stimulus controller for ${identifier}. if you use import.meta.glob, don't forget to enable the eager option to true`,
       );
     }
   });
+
+  if (app.debug) {
+    console.groupCollapsed("application register controllers from glob");
+    console.log("controllers", controllersAdded);
+    console.groupEnd();
+  }
 }
 
-export function registerController(app: Application, controllerInfos: StimulusControllerInfos) {
+export function registerController(app: Application, controllerInfos: StimulusControllerInfos, notify = true) {
   if (!controllerInfos.enabled) {
     return;
   }
@@ -110,5 +129,9 @@ export function registerController(app: Application, controllerInfos: StimulusCo
     app.register(controllerInfos.identifier, createLazyController(controllerInfos.controller));
   } else {
     app.register(controllerInfos.identifier, controllerInfos.controller);
+  }
+
+  if (app.debug && notify) {
+    console.log(`application register controller ${controllerInfos.identifier}`);
   }
 }
