+++ src/vite-plugin-symfony/src/stimulus/helpers/react/index.ts
@@ -1,23 +1,3 @@
-import { ImportedModule, ImportedModules, ReactModule } from "../types";
-
-let reactImportedModules: ImportedModules<ReactModule> = {};
-
-export function registerReactControllerComponents(
-  modules: ImportedModules<ReactModule>,
-  controllersDir = "./react/controllers",
-) {
-  reactImportedModules = { ...reactImportedModules, ...modules };
-
-  window.resolveReactComponent = (name: string): ImportedModule<ReactModule> => {
-    const reactModule =
-      reactImportedModules[`${controllersDir}/${name}.jsx`] || reactImportedModules[`${controllersDir}/${name}.tsx`];
-    if (typeof reactModule === "undefined") {
-      const possibleValues = Object.keys(reactImportedModules).map((key) =>
-        key.replace(`${controllersDir}/`, "").replace(".jsx", "").replace(".tsx", ""),
-      );
-      throw new Error(`React controller "${name}" does not exist. Possible values: ${possibleValues.join(", ")}`);
-    }
-
-    return reactModule;
-  };
-}
+export { ReactModule } from "./types";
+export * from "./util";
+export { default as ReactController } from "./render_controller";
