+++ src/vite-plugin-symfony/src/stimulus/helpers/react/types.ts
@@ -1,0 +1,5 @@
+import type { ComponentClass, FunctionComponent } from "react";
+export type ReactComponent = string | FunctionComponent<object> | ComponentClass<object, any>;
+export type ReactModule = {
+  default: ReactComponent;
+};
