+++ src/vite-plugin-symfony/src/stimulus/helpers/svelte/index.ts
@@ -1,22 +1,3 @@
-import { ImportedModule, ImportedModules, SvelteModule } from "../types";
-
-let svelteImportedModules: ImportedModules<SvelteModule> = {};
-
-export function registerSvelteControllerComponents(
-  modules: ImportedModules<SvelteModule>,
-  controllersDir = "./svelte/controllers",
-) {
-  svelteImportedModules = { ...svelteImportedModules, ...modules };
-
-  window.resolveSvelteComponent = (name: string): ImportedModule<SvelteModule> => {
-    const svelteModule = svelteImportedModules[`${controllersDir}/${name}.svelte`];
-    if (typeof svelteModule === "undefined") {
-      const possibleValues = Object.keys(svelteImportedModules).map((key) =>
-        key.replace(`${controllersDir}/`, "").replace(".svelte", ""),
-      );
-      throw new Error(`Svelte controller "${name}" does not exist. Possible values: ${possibleValues.join(", ")}`);
-    }
-
-    return svelteModule;
-  };
-}
+export { SvelteModule } from "./types";
+export * from "./util";
+export { default as SvelteController } from "./render_controller";
