+++ src/vite-plugin-symfony/src/stimulus/helpers/svelte/types.ts
@@ -1,0 +1,5 @@
+import type { SvelteComponent } from "svelte";
+
+export type SvelteModule = {
+  default: SvelteComponent;
+};
