+++ src/vite-plugin-symfony/src/stimulus/helpers/types.d.ts
@@ -1,23 +1,6 @@
-import type { ComponentClass, FunctionComponent } from "react";
-import { Component as VueComponent } from "vue";
-import type { SvelteComponent } from "svelte";
-
 export type LazyModule<M> = () => Promise<M>;
 export type ImportedModule<M> = M | LazyModule<M>;
 export type ImportedModules<M> = Record<string, ImportedModule<M>>;
-
-export type ReactComponent = string | FunctionComponent<object> | ComponentClass<object, any>;
-export type ReactModule = {
-  default: ReactComponent;
-};
-
-export type VueModule = {
-  default: VueComponent;
-};
-
-export type SvelteModule = {
-  default: SvelteComponent;
-};
 
 declare global {
   function resolveReactComponent(name: string): ImportedModule<ReactModule>;
