+++ src/vite-plugin-symfony/src/stimulus/helpers/vue/types.ts
@@ -1,0 +1,5 @@
+import { Component as VueComponent } from "vue";
+
+export type VueModule = {
+  default: VueComponent;
+};
