+++ src/vite-plugin-symfony/src/stimulus/util.ts
@@ -1,4 +1,4 @@
-export const CONTROLLER_FILENAME_REGEX = /^(?:.*?(controllers)\/|\.?\.\/)?(.+)\.[jt]sx?$/;
+export const CONTROLLER_FILENAME_REGEX = /^(?:.*?(controllers)\/|\.?\.\/)?(.+)\.[jt]sx?\b/;
 export const CONTROLLER_SUFFIX_REGEX = /^(.*)(?:[/_-]controller)$/;
 export function getStimulusControllerId(key: string, onlyControllersDir = false): string | null {
   const [, controllers, relativePath] = key.match(CONTROLLER_FILENAME_REGEX) || [];
